initialize pt_stack

+++ svm/svm.h
@@ -108,7 +108,7 @@ struct SVM_VirtualMachine_tag {
     uint8_t *stack_value_type;
     SVM_Value *stack;
     uint32_t pt_stack_size;  //
-    uint8_t *pt_stack;       //
+    size_t *pt_stack;        //
     size_t pt_stack_count;   //
     uint32_t pc;
     uint32_t sp;
